[RPC tests] Print verbose logging for TensorPipe
Some of our tests are flaky but when they fail they print very little information which makes it hard to diagnose the issue. One very useful piece of information which we'd love to have for the TensorPipe agent is its verbose logs. They are divided into many levels and printing all of them is probably quite impractical as that would create a bloat in the logs and a slowdown of the tests. A hopefully decent compromise is to only print the first level, which gives details about the interface between TensorPipe and PyTorch: which methods of TensorPipe are called by PyTorch, and which callbacks are invoked. This should hopefully not be too big, and would at least help "bisect" the issue into whether it's in TensorPipe or in the agent.

Differential Revision: [D23963480](https://our.internmc.facebook.com/intern/diff/D23963480/)

ghstack-source-id: 113026662
Pull Request resolved: https://github.com/pytorch/pytorch/pull/45440